Replace Integer transaction ID with Long
Redis returns a Long type integer (8 byte) on the INCR command.

+++ src/main/java/dev/unnm3d/rediseconomy/currency/Currency.java
@@ -511,7 +511,7 @@ public class Currency implements Economy {
      * @param transaction   The transaction to revert
      * @return The transaction id that reverted the initial transaction
      */
-    public CompletionStage<Integer> revertTransaction(int transactionId, @NotNull Transaction transaction) {
+    public CompletionStage<Long> revertTransaction(int transactionId, @NotNull Transaction transaction) {
         String ownerName = transaction.getAccountIdentifier().isPlayer() ?//If the sender is a player
                 currenciesManager.getUsernameFromUUIDCache(transaction.getAccountIdentifier().getUUID()) : //Get the username from the cache (with server uuid translation)
                 transaction.getAccountIdentifier().toString(); //Else, it's a bank, so we get the bank id
